Refactor projects to use hooks pattern
